refactor(application): reorganize CreateCountry command files into Create subfolder
- Moved CreateCountryCommand, CreateCountryCommandValidator, and CreateCountryHandler into a dedicated 'Create' subfolder under Commands for better structure and clarity
- Updated namespaces to AirlineBookingSystem.Application.CQRS.Countries.Commands.Create accordingly
- Adjusted unit tests to use the new namespace for CreateCountryHandler and related classes
